+++ public/app-config.js
@@ -1,0 +1,2 @@
+/* global window */
+window.__YANKI_CONFIG__ = {};